requirements.txt
alembic==1.2.1
click
Flask==1.1.1
Flask-Cors==3.0.8
numpy
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
PyMySQL==0.9.3
requests==2.22.0
scikit-image
SQLAlchemy==1.3.10
twilio==6.32.0
pytest
factory-boy
